pyproject.toml
[tool.poetry]
name = "issuer"
version = "0.1.0"
description = "Identity Controller for Aries Aca-Py"
authors = ["monar24 <mona@oneof.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
adrf = "0.1.1"
aiohttp = "3.8.4"
aiohttp-apispec = "2.2.3"
aiohttp-cors = "0.7.0"
aioredis = "2.0.1"
aiosignal = "1.3.1"
apispec = "3.3.2"
aries-cloudagent = "0.9.0"
asgiref = "3.6.0"
async-timeout = "4.0.2"
attrs = "22.2.0"
autopep8 = "2.0.2"
base58 = "2.1.1"
boto3 = "1.26.131"
botocore = "1.29.131"
cachetools = "5.3.0"
certifi = "2022.12.7"
cffi = "1.15.1"
chardet = "4.0.0"
charset-normalizer = "3.1.0"
configargparse = "1.5.3"
decorator = "5.1.1"
deepmerge = "0.3.0"
django = "4.1.7"
django-cors-headers = "3.14.0"
djangorestframework = "3.14.0"
ecdsa = "0.16.1"
frozenlist = "1.3.3"
gunicorn = "20.1.0"
hvac= "0.11.2"
idna = "2.10"
inflection = "0.5.1"
jinja2 = "3.1.2"
jmespath = "1.0.1"
jsonpath-ng = "1.5.2"
lxml = "4.9.2"
markdown = "3.1.1"
markupsafe = "2.0.1"
marshmallow = "3.19.0"
msgpack = "1.0.5"
multidict = "6.0.4"
nest-asyncio = "1.5.6"
packaging = "23.1"
pillow = "9.4.0"
ply = "3.11"
postgres = "4.0"
prompt-toolkit = "2.0.10"
psycopg2-binary = "2.9.6"
psycopg2-pool = "1.1"
pycodestyle = "2.10.0"
pycparser = "2.21"
pydantic = "1.9.2"
pydid = "0.3.7"
pyhcl = "0.4.4"
pyjwt = "2.8.0"
pyld = "2.0.3"
pynacl = "1.5.0"
pyparsing = "3.0.9"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-multipart = "0.0.6"
python3-indy = "1.16.0.post286"
pytz = "2021.3"
pyyaml = "6.0.1"
qrcode = "6.1"
rlp = "1.2.0"
s3transfer = "0.6.1"
six = "1.16.0"
sqlparse = "0.4.3"
typing-extensions = "4.0.1"
unflatten = "0.1.1"
urllib3 = "1.26.15"
wcwidth = "0.2.6"
webargs = "5.5.3"
yarl = "1.8.2"
drf_yasg = "1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"